Add support for account creation.
Standardize error messages from the remctl backend to start with
"error: " in case the Registry was depending on that.

Note that examine is the one case where we want to allow any principal
and fix a few other issues with the design document, including the
ordering of creation and the lack of need for an explicit queue.