Significant improvements to Kerberos and remctl configure probes
Attempt to determine if the library directory for remctl (and the
directory for Kerberos and GSS-API libraries if krb5-config isn't
found) is lib32 or lib64 instead of lib and set LDFLAGS accordingly.
Based on an idea from the CMU Autoconf macros.

Add --with-remctl-include, --with-remctl-lib, --with-gssapi-include,
--with-gssapi-lib, --with-krb5-include, and --with-krb5-lib configure
options to allow more specific setting of paths if necessary.

Check at configure time that we can link with the remctl library we
found.