Redo how examine of arbitrary instances is handled
The previous attempt still didn't work because there was nothing in the
%CONFIG hash for the unknown instance.  Now, fall back to the '' instance
for configuration on how to handle unknown instances and document that
fallback.

Improve the conversion of Kerberos v5 principal names to Kerberos v4
principal names for examine against an AFS kaserver, although it's
still not as good as calling the Kerberos library routine would be.