Added translator of Heimdal output to MIT get output
Heimdal uses a different syntax for the output of its get command than
MIT's getprinc.  The output from the examine command was originally based
upon the Heimdal::Kadm5 'dump' command, which models its output off of
Heimdal's get command.  Because downstream apps might be taking the output
of kadmin-backend examine, we replicate the Heimdal::Kadm5 code to
construct the get output here, but modelled to replicate MIT instead.  In
the long term, we would love to verify and have any applications using
the output fixed so that we aren't taking the arbitrary workaround.