Added ability to ignore enable for locked accounts
Added a new configuration option for each instance type, locked.  This is
an array containing a command and all arguments to it, which is used to
determine if a principal is marked to not be re-enabled via kadmin-backend.
If no command is set, none is checked.  The command is given the instance
name, and if the command returns 0, then the instance is locked and may
not be re-enabled via this interface.  This is done to support accounts
that may be locked for policy reasons, where we must go through extra steps
befor re-enabling the command.