Improve principal creation attributes for Heimdal
In the Heimdal backend, don't set KADM5_POLICY_NORMAL_MASK or
KADM5_POLICY_CLR as attributes when creating a new principal.  These
are not valid attribute values and end up setting or clearing large
numbers of other attributes.

In the Heimdal backend, don't unconditionally set the preauth required
attribute on newly created principals.  This should be handled using
the "default" principal in Heimdal to configure the desired default
principal lifetime and attributes.