Fixed to kadmin_create default attributes in kadmin-backend-heim
The code for creating a Kerberos principal was trying to get the default
set of attributes and then modify them before fully creating the account,
in order to properly create a disabled account.  However, the default
attributes seem to not be created until the principal is fully created,
so we actually lose the default attributes by doing so.  As a fix, hand
the routine our own default set of attributes.  This isn't optimal, but
it's better than nothing.