In Heimdal version, do password strength check with IPC::Run
Something about the workaround code to suppress the stderr result
from Heimdal's libraries causes STDERR handling to get messed up
in Perl.  Since the password strength checking program returns its
error on stderr, this is a problem.  IPC::Run works properly and is
much more succinct, so switch to it.