Set disallow-svr on all newly-created principals
Set the disallow-svr flag on all newly-created principals.  This
prohibits obtaining service tickets for the principal, which provides
some hardening against brute force attacks.  Since the create command
is designed for creation of user principals, not service principals,
and use of service tickets for user principals is quite obscure and
rare in Kerberos, this seems like a better default.